+++ src/cl/plan.cpp
@@ -12,15 +12,13 @@
 
 namespace bbfft {
 
-auto make_plan(configuration const &cfg, cl_command_queue queue, jit_cache *cache)
-    -> plan<cl_event> {
-    return plan<cl_event>(select_fft_algorithm<cl::api>(cfg, cl::api(queue), cache));
+auto make_plan(configuration const &cfg, cl_command_queue queue, jit_cache *cache) -> opencl_plan {
+    return opencl_plan(select_fft_algorithm<cl::api>(cfg, cl::api(queue), cache));
 }
 
 auto make_plan(configuration const &cfg, cl_command_queue queue, cl_context context,
-               cl_device_id device, jit_cache *cache) -> plan<cl_event> {
-    return plan<cl_event>(
-        select_fft_algorithm<cl::api>(cfg, cl::api(queue, context, device), cache));
+               cl_device_id device, jit_cache *cache) -> opencl_plan {
+    return opencl_plan(select_fft_algorithm<cl::api>(cfg, cl::api(queue, context, device), cache));
 }
 
 } // namespace bbfft