+++ src/sycl/api.cpp
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 #include "api.hpp"
-#include "build_wrapper.hpp"
 
 #include "bbfft/detail/compiler_options.hpp"
 #include "bbfft/sycl/device.hpp"
@@ -29,7 +28,6 @@ auto api::build_module(std::string const &source) -> shared_handle<module_handle
     return ::bbfft::sycl::make_shared_handle(
         ::bbfft::sycl::build_native_module(source, context_, device_, detail::compiler_options),
         queue_.get_backend());
-    ;
 }
 auto api::make_kernel_bundle(module_handle_t mod) -> kernel_bundle_type {
     return ::bbfft::sycl::make_kernel_bundle(mod, true, context_);