+++ src/sycl/plan.cpp
@@ -13,13 +13,13 @@
 
 namespace bbfft {
 
-auto make_plan(configuration const &cfg, ::sycl::queue q, jit_cache *cache) -> plan<::sycl::event> {
+auto make_plan(configuration const &cfg, ::sycl::queue q, jit_cache *cache) -> sycl_plan {
     return make_plan(cfg, q, q.get_context(), q.get_device(), cache);
 }
 
 auto make_plan(configuration const &cfg, ::sycl::queue q, ::sycl::context c, ::sycl::device d,
-               jit_cache *cache) -> plan<::sycl::event> {
-    return plan<::sycl::event>(select_fft_algorithm<sycl::api>(
+               jit_cache *cache) -> sycl_plan {
+    return sycl_plan(select_fft_algorithm<sycl::api>(
         cfg, sycl::api(std::move(q), std::move(c), std::move(d)), cache));
 }